feat(platform): extract copyToClipboad into reusable util
