+++ apps/platform/src/components/dashboard/project/projectCard/index.tsx
@@ -1,6 +1,5 @@
 'use client'
 import Link from 'next/link'
-import { toast } from 'sonner'
 import Avvvatars from 'avvvatars-react'
 import {
   SecretSVG,
@@ -24,6 +23,7 @@ import {
   editProjectOpenAtom,
   selectedProjectAtom
 } from '@/store'
+import { copyToClipboard } from '@/lib/cliboard'
 
 interface ProjectCardProps {
   project: ProjectWithCount
@@ -46,38 +46,6 @@ export default function ProjectCard({
   const setIsEditProjectSheetOpen = useSetAtom(editProjectOpenAtom)
   const setIsDeleteProjectOpen = useSetAtom(deleteProjectOpenAtom)
   const setSelectedProject = useSetAtom(selectedProjectAtom)
-
-  const copyToClipboard = (): void => {
-    // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition -- navigator.clipboard is checked
-    if (navigator.clipboard) {
-      navigator.clipboard
-        .writeText(`${window.location.origin}/project/${slug}`)
-        .then(() => {
-          toast.success('Link has been copied to clipboard.')
-        })
-        .catch((error) => {
-          // eslint-disable-next-line no-console -- console.error is used for debugging
-          console.error('Error copying text: ', error)
-        })
-    } else {
-      // Fallback for browsers that don't support the Clipboard API
-      // eslint-disable-next-line no-console -- console.log is used for debugging
-      console.log('Clipboard API not supported')
-
-      const textarea = document.createElement('textarea')
-      textarea.value = `${window.location.origin}/project/${slug}`
-      document.body.appendChild(textarea)
-      textarea.select()
-      try {
-        document.execCommand('copy')
-        toast.success('Link has been copied to clipboard.')
-      } catch (error) {
-        // eslint-disable-next-line no-console -- console.error is used for debugging
-        console.error('Error copying text: ', error)
-      }
-      document.body.removeChild(textarea)
-    }
-  }
 
   const handleEditProject = () => {
     setSelectedProject(project)
@@ -152,7 +120,7 @@ export default function ProjectCard({
         <ContextMenuItem
           inset
           onClick={() => {
-            copyToClipboard()
+            copyToClipboard(`${window.location.origin}/project/${slug}`)
           }}
         >
           Copy link
